latest

+++ src/kdseg/ds/Dot.java
@@ -4,9 +4,9 @@ public class Dot {
 	public Double lat;
 	public Double lng;
 	public TimeSegment seg;
-	public Double userkey;
+	public Integer userkey;
 	
-	public Dot(Double lat,Double lng,TimeSegment seg,Double userkey){
+	public Dot(Double lat,Double lng,TimeSegment seg,Integer userkey){
 		this.lat=lat;this.lng=lng;this.seg=seg;this.userkey=userkey;
 	}
 }
